requirements.txt
elasticsearch[async]>=8.6.0
python-decouple>=3.6